🚧 add plugin system : tidy up
